Liquid tags: Move tests in their own folder
