+++ liquid_tags/tests/test_generation.py
@@ -11,7 +11,7 @@ import pytest
 from pelican import Pelican
 from pelican.settings import read_settings
 
-from .notebook import IPYTHON_VERSION
+from ..notebook import IPYTHON_VERSION
 
 PLUGIN_DIR = os.path.dirname(__file__)
 TEST_DATA_DIR = os.path.join(PLUGIN_DIR, 'test_data')
